+++ src/events/DefineUserSettingEvent.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace wsydney76\contentoverview\events;
+
+use yii\base\Event;
+
+class DefineUserSettingEvent extends Event
+{
+    public string $key;
+    public mixed $value;
+}